+++ src/apps/ENERGY.hpp
@@ -10,59 +10,59 @@
 /// ENERGY kernel reference implementation:
 ///
 /// for (Index_type i = ibegin; i < iend; ++i ) {
-///   e_new[i] = e_old[i] - 0.5 * delvc[i] *
-///              (p_old[i] + q_old[i]) + 0.5 * work[i];
+///   e_new[i] = e_old[i] - 0.5f * delvc[i] *
+///              (p_old[i] + q_old[i]) + 0.5f * work[i];
 /// }
 ///
 /// for (Index_type i = ibegin; i < iend; ++i ) {
-///   if ( delvc[i] > 0.0 ) {
-///      q_new[i] = 0.0 ;
+///   if ( delvc[i] > 0.0f ) {
+///      q_new[i] = 0.0f ;
 ///   }
 ///   else {
-///      Real_type vhalf = 1.0 / (1.0 + compHalfStep[i]) ;
+///      Real_type vhalf = 1.0f / (1.0f + compHalfStep[i]) ;
 ///      Real_type ssc = ( pbvc[i] * e_new[i]
 ///         + vhalf * vhalf * bvc[i] * pHalfStep[i] ) / rho0 ;
-///      if ( ssc <= 0.1111111e-36 ) {
-///         ssc = 0.3333333e-18 ;
+///      if ( ssc <= 0.1111111e-36f ) {
+///         ssc = 0.3333333e-18f ;
 ///      } else {
-///         ssc = sqrt(ssc) ;
+///         ssc = sqrtf(ssc) ;
 ///      }
 ///      q_new[i] = (ssc*ql_old[i] + qq_old[i]) ;
 ///   }
 /// }
 ///
 /// for (Index_type i = ibegin; i < iend; ++i ) {
-///   e_new[i] = e_new[i] + 0.5 * delvc[i]
-///              * ( 3.0*(p_old[i] + q_old[i])
-///                  - 4.0*(pHalfStep[i] + q_new[i])) ;
+///   e_new[i] = e_new[i] + 0.5f * delvc[i]
+///              * ( 3.0f*(p_old[i] + q_old[i])
+///                  - 4.0f*(pHalfStep[i] + q_new[i])) ;
 /// }
 ///
 /// for (Index_type i = ibegin; i < iend; ++i ) {
-///   e_new[i] += 0.5 * work[i];
-///   if ( fabs(e_new[i]) < e_cut ) { e_new[i] = 0.0  ; }
+///   e_new[i] += 0.5f * work[i];
+///   if ( fabs(e_new[i]) < e_cut ) { e_new[i] = 0.0f  ; }
 ///   if ( e_new[i]  < emin ) { e_new[i] = emin ; }
 /// }
 ///
 /// for (Index_type i = ibegin; i < iend; ++i ) {
 ///   Real_type q_tilde ;
-///   if (delvc[i] > 0.0) {
+///   if (delvc[i] > 0.0f) {
 ///      q_tilde = 0. ;
 ///   }
 ///   else {
 ///      Real_type ssc = ( pbvc[i] * e_new[i]
 ///          + vnewc[i] * vnewc[i] * bvc[i] * p_new[i] ) / rho0 ;
-///      if ( ssc <= 0.1111111e-36 ) {
-///         ssc = 0.3333333e-18 ;
+///      if ( ssc <= 0.1111111e-36f ) {
+///         ssc = 0.3333333e-18f ;
 ///      } else {
-///         ssc = sqrt(ssc) ;
+///         ssc = sqrtf(ssc) ;
 ///      }
 ///      q_tilde = (ssc*ql_old[i] + qq_old[i]) ;
 ///   }
-///   e_new[i] = e_new[i] - ( 7.0*(p_old[i] + q_old[i])
-///                          - 8.0*(pHalfStep[i] + q_new[i])
-///                          + (p_new[i] + q_tilde)) * delvc[i] / 6.0 ;
+///   e_new[i] = e_new[i] - ( 7.0f*(p_old[i] + q_old[i])
+///                          - 8.0f*(pHalfStep[i] + q_new[i])
+///                          + (p_new[i] + q_tilde)) * delvc[i] / 6.0f ;
 ///   if ( fabs(e_new[i]) < e_cut ) {
-///      e_new[i] = 0.0  ;
+///      e_new[i] = 0.0f  ;
 ///   }
 ///   if ( e_new[i]  < emin ) {
 ///      e_new[i] = emin ;
@@ -70,16 +70,16 @@
 /// }
 ///
 /// for (Index_type i = ibegin; i < iend; ++i ) {
-///   if ( delvc[i] <= 0.0 ) {
+///   if ( delvc[i] <= 0.0f ) {
 ///      Real_type ssc = ( pbvc[i] * e_new[i]
 ///              + vnewc[i] * vnewc[i] * bvc[i] * p_new[i] ) / rho0 ;
-///      if ( ssc <= 0.1111111e-36 ) {
-///         ssc = 0.3333333e-18 ;
+///      if ( ssc <= 0.1111111e-36f ) {
+///         ssc = 0.3333333e-18f ;
 ///      } else {
-///         ssc = sqrt(ssc) ;
+///         ssc = sqrtf(ssc) ;
 ///      }
 ///      q_new[i] = (ssc*ql_old[i] + qq_old[i]) ;
-///      if (fabs(q_new[i]) < q_cut) q_new[i] = 0.0 ;
+///      if (fabs(q_new[i]) < q_cut) q_new[i] = 0.0f ;
 ///   }
 /// }
 ///
@@ -109,71 +109,71 @@
   const Real_type q_cut = m_q_cut;
 
 #define ENERGY_BODY1 \
-  e_new[i] = e_old[i] - 0.5 * delvc[i] * \
-             (p_old[i] + q_old[i]) + 0.5 * work[i];
+  e_new[i] = e_old[i] - 0.5f * delvc[i] * \
+             (p_old[i] + q_old[i]) + 0.5f * work[i];
 
 #define ENERGY_BODY2 \
-  if ( delvc[i] > 0.0 ) { \
-     q_new[i] = 0.0 ; \
+  if ( delvc[i] > 0.0f ) { \
+     q_new[i] = 0.0f ; \
   } \
   else { \
-     Real_type vhalf = 1.0 / (1.0 + compHalfStep[i]) ; \
+     Real_type vhalf = 1.0f / (1.0f + compHalfStep[i]) ; \
      Real_type ssc = ( pbvc[i] * e_new[i] \
         + vhalf * vhalf * bvc[i] * pHalfStep[i] ) / rho0 ; \
-     if ( ssc <= 0.1111111e-36 ) { \
-        ssc = 0.3333333e-18 ; \
+     if ( ssc <= 0.1111111e-36f ) { \
+        ssc = 0.3333333e-18f ; \
      } else { \
-        ssc = sqrt(ssc) ; \
+        ssc = sqrtf(ssc) ; \
      } \
      q_new[i] = (ssc*ql_old[i] + qq_old[i]) ; \
   }
 
 #define ENERGY_BODY3 \
-  e_new[i] = e_new[i] + 0.5 * delvc[i] \
-             * ( 3.0*(p_old[i] + q_old[i]) \
-                 - 4.0*(pHalfStep[i] + q_new[i])) ;
+  e_new[i] = e_new[i] + 0.5f * delvc[i] \
+             * ( 3.0f*(p_old[i] + q_old[i]) \
+                 - 4.0f*(pHalfStep[i] + q_new[i])) ;
 
 #define ENERGY_BODY4 \
-  e_new[i] += 0.5 * work[i]; \
-  if ( fabs(e_new[i]) < e_cut ) { e_new[i] = 0.0  ; } \
+  e_new[i] += 0.5f * work[i]; \
+  if ( fabs(e_new[i]) < e_cut ) { e_new[i] = 0.0f  ; } \
   if ( e_new[i]  < emin ) { e_new[i] = emin ; }
 
 #define ENERGY_BODY5 \
   Real_type q_tilde ; \
-  if (delvc[i] > 0.0) { \
+  if (delvc[i] > 0.0f) { \
      q_tilde = 0. ; \
   } \
   else { \
      Real_type ssc = ( pbvc[i] * e_new[i] \
          + vnewc[i] * vnewc[i] * bvc[i] * p_new[i] ) / rho0 ; \
-     if ( ssc <= 0.1111111e-36 ) { \
-        ssc = 0.3333333e-18 ; \
+     if ( ssc <= 0.1111111e-36f ) { \
+        ssc = 0.3333333e-18f ; \
      } else { \
-        ssc = sqrt(ssc) ; \
+        ssc = sqrtf(ssc) ; \
      } \
      q_tilde = (ssc*ql_old[i] + qq_old[i]) ; \
   } \
-  e_new[i] = e_new[i] - ( 7.0*(p_old[i] + q_old[i]) \
-                         - 8.0*(pHalfStep[i] + q_new[i]) \
-                         + (p_new[i] + q_tilde)) * delvc[i] / 6.0 ; \
+  e_new[i] = e_new[i] - ( 7.0f*(p_old[i] + q_old[i]) \
+                         - 8.0f*(pHalfStep[i] + q_new[i]) \
+                         + (p_new[i] + q_tilde)) * delvc[i] / 6.0f ; \
   if ( fabs(e_new[i]) < e_cut ) { \
-     e_new[i] = 0.0  ; \
+     e_new[i] = 0.0f  ; \
   } \
   if ( e_new[i]  < emin ) { \
      e_new[i] = emin ; \
   }
 
 #define ENERGY_BODY6 \
-  if ( delvc[i] <= 0.0 ) { \
+  if ( delvc[i] <= 0.0f ) { \
      Real_type ssc = ( pbvc[i] * e_new[i] \
              + vnewc[i] * vnewc[i] * bvc[i] * p_new[i] ) / rho0 ; \
-     if ( ssc <= 0.1111111e-36 ) { \
-        ssc = 0.3333333e-18 ; \
+     if ( ssc <= 0.1111111e-36f ) { \
+        ssc = 0.3333333e-18f ; \
      } else { \
-        ssc = sqrt(ssc) ; \
+        ssc = sqrtf(ssc) ; \
      } \
      q_new[i] = (ssc*ql_old[i] + qq_old[i]) ; \
-     if (fabs(q_new[i]) < q_cut) q_new[i] = 0.0 ; \
+     if (fabs(q_new[i]) < q_cut) q_new[i] = 0.0f ; \
   }
